Added isFalseyString
Hi,

I’d like to propose a new function isFalseyString. This function
determines if a given value is a string representation of a falsey
type. This is exceptionally useful when api data is stored either in
the dom or in local storage as a string and it’s falsey type is lost.